+++ docs/reference/bigframes.bigquery/index.rst
@@ -1,0 +1,9 @@
+
+===========================
+BigQuery Built-in Functions
+===========================
+
+.. automodule:: bigframes.bigquery
+    :members:
+    :inherited-members:
+    :undoc-members:
