+++ docs/reference/bigframes/index.rst
@@ -1,4 +1,5 @@
 
+============
 Core objects
 ============
